// src/app/core/models/may-bien-ap.ts
export class MayBienApList {
  MAPMIS: string = '';
  TENCONGTY: string = '';
  TRUYENTAIDIEN: string = '';
  TENMBA: string = '';
  TENTRAM: string = '';
}

export class MayBienApDetail {
  Mapmis: string | null | undefined;
  Madvql: string | null | undefined;
  Tencongty: string | null | undefined;
  Truyentaidien: string | null | undefined;
  Tenmba: string | null | undefined;
  Sohieu: string | null | undefined;
  Sohuu: string | null | undefined;
  Ngaylapdat: Date | null | undefined;
  Ngayvh: Date | null | undefined;
  Thuoctram: string | null | undefined;
  Tentram: string | null | undefined;
  Hangsx: string | null | undefined;
  Soserial: string | null | undefined;
  Sohieubanve: string | null | undefined;
  Sododanhso: string | null | undefined;
  Mach: string | null | undefined;
  Cottenhienthi: string | null | undefined;
  Dahienthitrensd: string | null | undefined;
  Hienthiten: string | null | undefined;
  Hoatdong: string | null | undefined;
  Tthientai: string | null | undefined;
  Jsongeo: string | null | undefined;
  Maudong: string | null | undefined;
  Maucat: string | null | undefined;
  Daunoi: string | null | undefined;
  Ghichu: string | null | undefined;
}